Add files via upload
This is a big project I worked on for a real estate agents website. I used PHRETS to help me connect to their MLS feed (New England Real Estate Network - MLS) and used PHP to write code and connect to their server, create a database for the active MLS listings (ALL active listings for New England), and periodically check the MLS feed to update their SQL database with new listings and/or to remove inactive listings.